[강운덕] [LUCYSUS-1744] test가 깨지는 testcase 수정.
git-svn-id: http://svn.bds.nhncorp.com/pe/hippo-commons/trunk@1639 84d0f5b1-2673-498c-a247-62c4ff18d310

+++ src/test/java/com/profiler/common/util/AnnotationTranscoderTest.java
@@ -60,7 +60,7 @@ public class AnnotationTranscoderTest {
     private void typeCode(Object value) {
         AnnotationTranscoder transcoder = new AnnotationTranscoder();
 
-        int typeCode = transcoder.getTypeCode(value);
+        byte typeCode = transcoder.getTypeCode(value);
         byte[] bytes = transcoder.encode(value, typeCode);
         Object decode = transcoder.decode(typeCode, bytes);
 
@@ -70,7 +70,7 @@ public class AnnotationTranscoderTest {
     private void typeUnsupportCode(Object value) {
         AnnotationTranscoder transcoder = new AnnotationTranscoder();
 
-        int typeCode = transcoder.getTypeCode(value);
+        byte typeCode = transcoder.getTypeCode(value);
         byte[] bytes = transcoder.encode(value, typeCode);
         Object decode = transcoder.decode(typeCode, bytes);
 
@@ -80,7 +80,7 @@ public class AnnotationTranscoderTest {
     private void typeBinaryCode(byte[] value) {
         AnnotationTranscoder transcoder = new AnnotationTranscoder();
 
-        int typeCode = transcoder.getTypeCode(value);
+        byte typeCode = transcoder.getTypeCode(value);
         byte[] bytes = transcoder.encode(value, typeCode);
         Object decode = transcoder.decode(typeCode, bytes);
